requirements.txt
openai
requests
numpy
xmltodict
beautifulsoup4
python-dotenv
sentence-transformers
langchain
scikit-learn

